// frontend/todo-frontend/src/app/model/profiles.ts
export interface Profiles {
  id: number;

  fullName?: string;
  dateOfBirth?: string;

  street?: string;
  city?: string;
  state?: string;
  country?: string;
  postalCode?: string;
  phoneNumber?: string;

  title?: string;
  experience?: number;
  bio?: string;
  skills?: string;

  photoId?: string;
}